+++ shared/src/main/java/com/skopware/vdjvis/api/entities/StatusBayar.java
@@ -5,6 +5,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class StatusBayar {
     @JsonProperty public int status;
     @JsonProperty public String strStatus;
-    @JsonProperty public long countBulan;
-    @JsonProperty public long nominal;
+    @JsonProperty public int countBulan;
+    @JsonProperty public int nominal;
 }
